Remove sorter to services folder
