+++ src/constants/Group.js
@@ -10,13 +10,3 @@ export const groupMemberType = {
   PENDING_MEMBER: 3,
   NON_MEMBER: 4
 };
-
-export const groupMemberSorter = (x, y) => {
-  const groupRoles = ["Owner", "Admin", "Member"];
-
-  if (x.groupRole === y.groupRole) {
-    return x.username - y.username;
-  }
-
-  return groupRoles.indexOf(x.groupRole) - groupRoles.indexOf(y.groupRole);
-};
